+++ test/test_trezor.py
@@ -66,8 +66,9 @@ class TrezorEmulator(DeviceEmulator):
         return client
 
     def stop(self):
-        os.killpg(os.getpgid(self.emulator_proc.pid), signal.SIGTERM)
-        os.waitpid(self.emulator_proc.pid, 0)
+        if self.emulator_proc.poll() is None:
+            os.killpg(os.getpgid(self.emulator_proc.pid), signal.SIGTERM)
+            os.waitpid(self.emulator_proc.pid, 0)
 
         # Clean up emulator image
         if self.model_t:
